Allow Formchecker to check more than 1 thing at a time
The FormChecker is currently limiting the number of checks to one at a
time. While for most forms this is fine for the matrix auth and the
role-strategy this makes their config pages very slow to load. When 300
users are configured against an Active Directory it takes for me well
over 10 minutes until for each sid it is decided if it is a user or
group or doesn't exist.
Changing this needs to handled with care by those plugins. Browsers
limit the number of parallel connections one can open to the same host
usually with http1.1
But with http2 one can see big performance improvements. Using a
value of 50 and a role with 300 users the loading of user information
dropped from over 10 min to about 1 min for me.