[JENKINS-69650] Make Security2779Test failure easier to diagnose
Apache Maven surefire plugin does not seem to fully support parameterized
JUnit 4 tests.  See https://github.com/junit-team/junit5/issues/990 for
discussion of the issue.

The surefire plugin acts as though the Security2779Test is being run
multiple times with the same argument and treats a failure of one argument
as a "flaky test".  In this case, the test is not flaky as far as I can
tell, it is showing a real failure when the argument is "#icon-panel svg".

The test passes consistently when the argument is "#link-panel a".

The consistent failures that I've seen were on Red Hat Enterprise Linux
8.6 with Eclipse Temurin Java 11.0.16.1.  The same failure was seen in
the Jenkins 2.369 weekly build.

This commit does not fix the failure, it just makes it easier to see
that there is a failure.