create Badge class
Badge class takes message, tooltip and severity

Do not create badge for NodesLink. It will not work with the
agent-maintenance plugin as this configures as well a retention strategy
that can't be queried here. To get the reliably working Demand and
SimpleScheduled would need some reworking at least.

The plugin link badge now changes color when there is an update that
fixes a security vulnerability.