Avoid URICheckEncodingMonitor empty value NPE
https://community.jenkins.io/t/reverse-proxy-test-does-not-handle-url-correctly/1294/9
reports that an empty value returned to the URI encoding check monitor
will report a null pointer exception.  An empty value is unexpected in
that context, but it is easy to check for null and avoid the null
pointer exception.

No Jira ticket was opened for the issue.  I can open a Jira ticket if
that is required.