Remove 'prepare changelogs' from release team
The release lead uses the release checklist that is linked in a later
sentence.  The release checklist guides the release lead to assign and
monitor tasks include changelog preparation, upgrade guide creation,
backports, release candidate delivery, and more.

I don't think we need to specifically mention "prepare changelog" in
this document when the changelog and upgrade guide has been created by
a Docs team member for the last 18 months or more.

I think we should leave initiate release and perform backports because
those are common tasks for the release lead and they are high
visibility tasks.