[JENKINS-70599] restore installNecessaryPlugins redirect destination
The installNecessaryPlugins API end point previously redirected to the
"/updateCenter" URL.  When the side bar was added to the plugin
manager, the redirect was changed to "/updates".  Unfortunately, there
is no page at "/updates", while there is a page at "/updateCenter".

This restores the redirect for the installNecessaryPlugins API call.

If the redirect is not restored, calls to the API that follow the redirect
fail with a page not found (HTTP 404).

I did not find any callers to that API from inside Jenkins core.

Change introduced in 6db88c3f5ba82ecc68af2bced3a568058cf80e2b