One copy button per region
Creates a heading over the region to be copied and create a set of
instructions for Unix agents with a different set of instructions
for Windows.  The `curl.exe` call must be different on Windows due to
PowerShell having a `curl` cmdlet that is not compatible with command line
`curl.exe`.

Passed my interactive testing with a session in a secure connection.