Adjust test timeouts based on actual test results
Tests on ci.jenkins.io showed that installNecessaryPluginsTest completes
as follows:

* 9.4 seconds Windows JDK 17
* 0.8 seconds Linux JDK 11
* 1.0 seconds Linux JDK 17
* 1.1 seconds Linux JDK 19

With those timings, it seems reasonable to set the crumb request to
timeout in not more than 7 seconds and the plugin installation request
to timeout in not more than 20 seconds.  That will waste a little less
time in execution if there is a problem and still leaves room for low
performance systems.