[JENKINS-70599] restore installNecessaryPlugins redirect destination (#7653)
* Test redirect from installNecessaryPlugins

Jenkins LTS versions prior to 2.375.1 redirected to an existing URL.
Jenkins 2.375.1 - 2.375.3 redirect to a non-existent URL.

Test that the old behavior is restored.

* [JENKINS-70599] restore installNecessaryPlugins redirect destination

The installNecessaryPlugins API end point previously redirected to the
"/updateCenter" URL.  When the side bar was added to the plugin
manager, the redirect was changed to "/updates".  Unfortunately, there
is no page at "/updates", while there is a page at "/updateCenter".

This restores the redirect for the installNecessaryPlugins API call.

If the redirect is not restored, calls to the API that follow the redirect
fail with a page not found (HTTP 404).

I did not find any callers to that API from inside Jenkins core.

Change introduced in 6db88c3f5ba82ecc68af2bced3a568058cf80e2b

* Use default HttpClient version

The HttpClient documentation notes that if the default is selected but
not supported, then the client will automatically downgrade to HTTP 1.1.
Browsers only support HTTP 2 over TLS, but that's not a requirement of
the HTTP 2 specification.

https://http2.github.io/faq/#does-http2-require-encryption

* Adjust test timeouts based on actual test results

Tests on ci.jenkins.io showed that installNecessaryPluginsTest completes
as follows:

* 9.4 seconds Windows JDK 17
* 0.8 seconds Linux JDK 11
* 1.0 seconds Linux JDK 17
* 1.1 seconds Linux JDK 19

With those timings, it seems reasonable to set the crumb request to
timeout in not more than 7 seconds and the plugin installation request
to timeout in not more than 20 seconds.  That will waste a little less
time in execution if there is a problem and still leaves room for low
performance systems.

* Redirect to newer pluginManager URL

(cherry picked from commit a7b36ef111cc61262a4c3619dff6b57ab49fe663)