[JENKINS-21052] Warn user that the copy button requires HTTPS (#7665)
Warn user that the copy button requires HTTPS

Users that run Jenkins over an unencrypted (HTTP) connection will see
the copy button, but it does not perform a copy.  Browsers will only
allow access to the clipboard if the page is in a secure context.

With this change, the text that is displayed when the user clicks the
copy button from an insecure context will alert the user that copy
requires a secure connection.

https://stackoverflow.com/questions/400212/how-do-i-copy-to-the-clipboard-in-javascript
describes the alternatives in more detail.

https://issues.jenkins.io/browse/JENKINS-21052 reported that the copy
button does not work.  I've confirmed that it works when using an HTTPS
connection.