create dedicated 404 error page
Currently Jenkins serves the error page of the underlying servlet
container when a page doesn't exist or the permissions deny the
discovery of the page.
The user experience will be better if a page is returned that will look
like all other pages including navigation back to the root. And also
add a hint that it might be a permission problem.