modal with title for legend
The modal feature is displaying to complete content inside a div. For
the legend which is rather big the content must be scrolled. At present
this lead to the content being displayed without any margins and the
scrollbar arrows at top and bottom are partially overlapping with the
rounded corners of the modal.

This change introduces a dedicated title for the modal. The title will
be fixed on top of the modal and only the content is scrollable. A
proper margin at the bottom ensures that the scrollbar is not
overlapping with the corners.