Disable spotless groovy formatting
There was an outage in eclipse infrastructure causing unnecessary build
issues.

The plugin was accessing
https://download.eclipse.org/eclipse/updates/4.26/ which was returning a
HTTP 502 error, and it was then throwing
org.tukaani.xz.XZFormatException: Input is not in the XZ format because
it was downloading an html file with the http 502 error.

The spotless plugin should not bring dependencies from 3rd party
websites that can't be mirrored in the jenkins infrastructure.

https://github.com/diffplug/spotless/issues/1697 is recording the
problem on spotless side, but until then, getting stable builds
outweights the benefit from formatting groovy files