Replace disconnect and system info symbols for agents (#8015)
* replace disconnect and system info symbols for agents

The symbol to disconnect was the same as that to delete the agent. Add a
new symbol that better fits for disconnecting (There is no fitting icon
at ionicons, so this is one from tabler-icons that I modified to have a
bigger viewport).
The symbol for System Information was still on old colored svg. Replace
this with information-circle from ionicons

* Make info theme-able

---------

Co-authored-by: Alexander Brandes <mc.cache@web.de>