Detect if getJSON does not throw an exception
The original test that was the basis for the new test does not detect
that the exception was not thrown.  The `assertThrows` method of JUnit
4 and JUnit 5 is a better choice for that type of assertion, because it
fails the test if the exception is not thrown.

Original test is updated to match current behavior.

Original test uses a hamcrest assertion to check the start of the
returned string.  Complete string includes a locale component and the
test does not need to assert the current locale.

New test is updated to match current behavior.