Evaluate if JVM is Java 9+ by inverting 'Java 8 or older' check
Given the fact that JVM versions were prefixed with '1.' until Java 9,
there are 2 helper methods regarding the Java version being used; 'Java 8 or older' and 'Java 9 or newer'.

Since the methods are mutually exclusive and are essentially an inverse of one another, the issue was that both methods get the runtime Java version and then check if the version begins with '1.'. But because of the inverse property, only one method needs to do this check, and the other call the first and invert the result.