Remove incorrect text from JENKINS_HOME help page (#8161)
Since the transition to systemd, we no longer read the contents of
/etc/sysconfig/jenkins on Red Hat Enterprise Linux systems.  The changes
that a user can make with systemd are either modifying the environment
variable or they are modifying a Java property.  Since those two cases
are already covered, let's not mention the different ways of starting
the controller.

We don't actively promote the use of alternate web containers and
setting the environment variable in the web container admin tool is
already covered by the simpler statement that the environment variable
needs to be set.

We absolutely should not suggest that they modify the contents of the
jenkins.war file.  That type of change will likely be erased on the
next upgrade.

https://github.com/jenkins-infra/jenkins.io/pull/6409 review process
showed this issue.