allow string for badge attribute of l:task
The documentation states that the badge attribute accepts a value as
argument. But in reality it requires an instance of
jenkins.management.Badge

The effect can be seen on the Design Library
https://weekly.ci.jenkins.io/design-library/Layouts/twoColumn
Where a String is passed but not shown.

This change allows to have both either a Badge object or something else
which is then just printed.