persist the disabling of stats when disabled by system property
we can no longer toggle the stats option when stats are disabled by
system property.
but this can lead to callers of the API to think that it is enabled.

so if stats are disabled by system property then we disable stats in
the config so users of the API behave appropriately