remove jcifs and j-interop
jinterop existed soley so we can install Jenkins as a service via the
Jenkins UI.

This functionality should not be used, and the windows installer should
be used instead.

Removed the jcifs library as it is outdated, no longer maintained and
was used only to support launching windows agents via DCOM which no
longer works with supported windows versions.