[JENKINS-72170] fix for nested-hetero-list entries
when a plugin defines the hetero-list elements on its own and still uses
an <input type="button"> element and an inner element is already using
a hetero-list with a <button> element, then the old input wasn't
converted to a button as the inner button was found first.
Avoid this by converting all inputs to buttons before actually looking
up the button.

Recreate the hetero-list.js file to avoid the "No such adjunct found"
warning message