do not specify anti-aliasing
specifying the anti-aliasing implies we know better than the browser (we
don't).  Specifiying this globally prevents the use of sub-pixel anti-aliasing
where it is available and the browsers text rendering engines are these
days pretty much fantastic that they should not need these hacks.

and for good measure - here is an article from 10 years ago
https://usabilitypost.com/2012/11/05/stop-fixing-font-smoothing/
and the mozilla doc saying do not use it on a public facing web site.
https://developer.mozilla.org/en-US/docs/Web/CSS/font-smooth