[JENKINS-72371] rework node monitor configuration (#8719)
* [JENKINS-72371] rework node monitor configuration

This PR reworks the way node monitors are configured. It ensures that
also monitors which are set to ignored, can be configured.
Previously, there was a checkbox before each monitor, where the
behaviour was not totally clear. It meant that the monitor is ignored
but still collecting data and not disabled as one might think. But when
the monitor was disabled any configuration was lost and default values
were used.

* improve description

* fix formatting

* add since