Use spotbugs plugin 4.8.2.0 with additional suppressions
https://github.com/jenkinsci/pom/pull/510 needs this change along with
checks of the other repositories that use the Jenkins parent pom.

https://github.com/jenkinsci/pom/pull/501#pullrequestreview-1718536231
recommends that the new spotbugs issues be either suppressed or resolved
in the upstream release of spotbugs.  This change suppresses the new
warning for primitive fields that are pu9blicly visible and the new
warning for a few cases where a field could be made static.

This change skips the warnings related to CT_CONSTRUCTOR_THROWS because
they are not relevant to Jenkins.

https://github.com/jenkinsci/plugin-pom/pull/869#issuecomment-1860918407
provides more details along with the discussion at
https://github.com/spotbugs/spotbugs/issues/2695