[JENKINS-69191] icon for agent not accepting tasks
This change adds another icon for the case when an agent is not
accepting tasks, e.g. because the retention strategy avoids this.
The makes it easier to decide whether an agent is in a non healthy
state.
A legend button is added below the computer overview table that shows an
explanation for the icons.