do not generate td when outside table for buildbutton
projectView renders a buildButton twice once inside a table for wide
screens and once outside a table for narrow or mobile screens with one
always hidden. But the buildButton always wraps everything in a `td`.
When projectView is now itself wrapped somewhere in a table (was done
in dashboard-view plugin) then the brwoser will move the `td` to the
outer table and it gets always shown and breaks the UI.