[JENKINS-72407] missing folder icon (#8872)
when a folder icon is provided and relies on the getImageOf method this
leads to a missing icon, e.g. gitlab branch source or when using a
custom folder icon
So fall back to the ballColorTd when we have no iconname which handles
all cases properly.

(cherry picked from commit 70f2237147f238adb54b21edf3e354fc672aeae3)