[JENKINS-72509] avoid broken log when deleting job/agent
when accessing the log of a running build and the job gets deleted or
someone removes read permissions (without discover) the log just appends
the output of the call to logText/progressiveHtml which is the 404
message from Jenkins.
The same happens when one accesses the log of an agent and the agent is
deleted.
This is due to only checking for 403 status and ignoring other status
codes from the 400 range.