fix margin in readonly mode
when readonly mode is enabled with systemread or extended read
permissions, some fields are wrapped in a `pre` tag. By default `pre`
has a bigger margin-bottom . But in readonly mode we already have a
margin in a wrapping div which leads to unnecessary large space after
those fields.
So we should remove the margin-bottom for this