[JENKINS-72679] avoid admin monitor popup makes buttons unusable (#8941)
avoid admin monitor makes buttons unusable

The div of the admin monitor pop-up is set to opacity 0 but keeps
z-index 1000 when closed. This makes the buttons, that are behind the
popup when shown unusable.
This change uses an animation that ensures the z-index is 0 after hiding
the popup and buttons are still usable

(cherry picked from commit a6423541f07292f8e5b74a9ce0e9a4d711eda0d9)