[JENKINS-72944] reduce spacing in buildhistory widget
the logic when name, details and badges arrangement is changed so that
the available space is better used.
The stop action is no longer treated as separate thing, instead it is just
like all other badges. The stop button will always be in the far right.
The width calculation now considers the actual width of the content and
not of the element they are in. When name and badges fit in one
row they will be joined.
badges are using a flex layout that should allow for better alignement
and will automatically wrap the elements though this only happen now
when the badges are in their own row and still to large

Add a resize handler so when the sidepanel switch to/from tablet mode or
size is changed in tablet mode then the table layout is recalculated.

[JENKINS-70246] the sidepanel is now fixed width when total width is
larger than 900px.