[JENKINS-72744] remove tooltip in widget refresh
Attach the tooltip to the parent element instead of the document body.
When a widget is refreshed, the elements are deleted. But the tooltip
stays when it is added to the body and thus will never be able to
receive an event when the mouse is moved out.
By attaching it to the parent, when the widget is refreshed and the
content is remobed this will automatically also remove the tooltip html
elements.