[JENKINS-53958] allow pipeline jobs to run when built-in is offline
When a pipeline starts building it creates an OneOffExecutor that takes
care of the pipeline execution on the built-in node.
The executor has some logic to prevent running things when an agent has
gone offline in the timeframe between assiging the task to the executor
and the executor actually starting running the task. But this logic
falsely leads to a termination of the executor for the pipeline job and
the attempts to restart the task fails as the task is no longer in the
queue.
Thas change tries to avoid this by ignoring the online state for the
built-in node as it will never be really offline (there is no channel
that can be closed). One can take it temporarily offline but this should
not prevent pipelines that do not explicitly make use of the built-in to
start running.