[JENKINS-73114] avoid conflict of css classes with bootstrap
jenkins and bootstrap both have definitions for alert and alert-warning,
alert-info, alert-danger. Bootstrap css definitions are included when
e.g. the warning-ng plugin is installed and a job is configured to scan
for warnings. The css classes from bootstrap are then chosen instead of
the ones for Jenkins for things like the admin monitors.
To make things umambiguous, add additional classes prefixed with
`jenkins-` and make use of them in code. Keep the old definitions for
backward compatibility with plugins.
Followup changes are needed in plugins (Mostly those that define admin
monitors) and the design-library