[JENKINS-21605] Further reduce number of recorded upstream causes (#9248)
* Fix non-determinism in nested cause tests

For deeplyNestedCauses, quickly triggering a jobs with no quiet period
runs the possibility that a job is triggered while it is still in the
queue. Triggering and waiting for the jobs to execute synchronously
avoids this and also simplifies the code.

The broadlyNesstedCauses test suffers from the same problem, but is
additionally not structured in the way the original issue describes to
reproduce the issue. Using the quiet period argument to scheduleBuild2,
we can reliably produce a graph where each build is triggered by the
other two jobs. An additional final job outside of the loop cuts off the
1 second quiet period from the last loop's iteration so that the test is
never waiting for the quiet period of any jobs even though half of the
schedule operations specify one.

* Re-work nested cause trimming algorithm