[JENKINS-73158] avoid jumping layout due to tooltips
The fix for JENKINS-72744 made the tooltip get appended to the parent
element. But this can also cause troubles in some cases.
To avoid this one can now decide whether the tooltip should be appended
to the body (the default) or to the parent.
Adjust all locations where a tooltip is displayed inside a widget to
append it to the parent. This works well with those tooltips and avoids
a jumping layout in the custom-folder-icon and potentially other places.