Revert part of "Remove dead JS code (#9136)"
https://issues.jenkins.io/browse/JENKINS-73166 notes that the credentials
code for certificates has a dependency on this JavaScript code.
Without this JavaScript, an attempt to add a new certificate credential
type using PKCS#12 will fail with an HTTP 404 error.

The replacement of YUI.log with console.log is not involved in the bug
report, so it is retained in the code change.

This reverts part of commit 932ab19c5448bb0afa46f60b93c245eb08e86b55.