move the add/edit description button to the app-bar (#9271)
* move the add/edit description button to the app-bar

So far rhe add/edit description button is rendered as button on its own
line to the far right. This leads to additional space the description or
the title and the following content.

This change moves the add/edit description button to the app-bar where
applicable (jobs, runs, computers, users). For views it stays as it is
as they don't have title/app-bar. One also can't just allow to change
the description via the configure page as the All View has no configure
page.

* make secondary button

* feedback

- fix permission in job page
- don't show the button for built-in node
- improve jelly doc