Fix incorrect assumptions in Test
Whilst the test was correct that RJR would load detached plugins it was
actually relying on it to do so to load instance-idenentity so that JNLP
(sic) agents would work.

if RJR is updated then this assumption no longer holds and the test
would time out waiting for the JNLP agent to connect.

instance-identity is already in the test classpath - and we are not
testing that this functionality works with no plugins or a limited set,
so no longer remove any test scoped plugin from the test.

ammends #6585