Update javadoc and other docs with dates of API changes (#9396)
Several of the changes are not detected by the Python script but were
detected by `git grep TODO | grep since` with review of the output.

The cases that are not detected by the Python script seem obscure enough
to not bother with additions to the script.  If they become common
practice, then we can extend the Python script.