Adapt to Secrets being sent as "" rather than null
values for SecretTextArea are now always sent.  Whilst this is a change
in behavior the only way once set to unset is to set the value to the
mpty string "" which would still produce a Secret

Additionally this matches the behaviour of f:password and f:textarea

Given both of the above this change in behaviour should not break any
production use cases.