Only skip tests on ci.jenkins.io master branch
Continues to run the tests on pull request branches, on developer
environments, and on other CI servers like release.ci.jenkins.io and
cert.ci.jenkins.io.

Run the test approximately 1 of 8 times on the ci.jenkins.io master branch
as well.  In the last 4 weeks, we've had 7.5 builds of the master branch
completed each week.  Running the tests 1 of 8 times should provide about
3 runs per week, since we run 3 configurations (Windows with Java 17,
Linux with Java 17, and Linux with Java 21).