[JENKINS-73404] SecretTextArea missing FormValidation support (#9450)
* [JENKINS-73404] unit test

* roundtrip as Secrets

* [JENKINS-73404] add validation for SecretTextArea

reworked the jelly and simplified the javascript.

the textarea is now always present in the DOM, so no need to remove
inputs to add a textarea.  Rather we just remove /hide the "hidden" attribute
from the elements we want to show/hide and give the textArea the
expected class and attributes for validation to occur

* remove unused data-prompt attribute

* Adapt getHiddenSecretValue to backend changes

* Adapt to Secrets being sent as "" rather than null

values for SecretTextArea are now always sent.  Whilst this is a change
in behavior the only way once set to unset is to set the value to the
mpty string "" which would still produce a Secret

Additionally this matches the behaviour of f:password and f:textarea

Given both of the above this change in behaviour should not break any
production use cases.

* Remove trailing spaces

---------

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>
(cherry picked from commit 81b90e4c118bcdc9b283037e819fb2d6217da18a)