Remove Jira issue from changelog summary in PR template (#9490)
When the Jira issue is included in the changelog summary, it is included
in the generated changelog as the initial text of the changelog entry.
The same issue link is then included by the changelog generator at the end
of the changelog description.  Remove the example that places the issue at
the start of the changelog text and explain that the issue link should be
elsewhere in the pull request so that the changelog generator can use it.

When the Jira issue is included in the changelog summary, the changelog
maintainers must remove it before the changelog is published.  It is
better to not include the Jira issue in the changelog summary so that
changelog maintainers do not need to remove it.

Remind submitters that they should not delete the upgrade guidelines
section because the presence of the upgrade guidelines section is used
by the changelog generator as the terminator for the proposed changelog
entries.