Use empty alt text when immediately followed by text
As noted by Daniel Beck, it does not help the user of the screen
reader for us to show the same fundamental text message twice.  The
image in these cases is a hint for the reader of the message text that
immediately follows the image.  Cluttering the screen reader content
with a second copy of the same message is not helpful.