Add end of life dates for Alpine 3.20, Ubuntu 24.04, Fedora 40
Also update the end of life dates for other operating systems based on
https://endoflife.date/ data.

Removes the 'start' entries because we are past the end of life date
of the RHEL 7 derivatives and the one other start date is not needed
because the default start date will cover it correctly.

Confirmed with a CentOS 8 computer that the end of life message
is displayed with the correct date for the CentOS 8 end of life.
The corrected end of life date has already passed, so there may be
Jenkins users that are surprised when their next upgrade reports that
they are running an unsupported operating system.